// src/services/openapi/models/FriendDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FriendDeviceDto } from './FriendDeviceDto';
export type FriendDto = {
    id: string;
    displayName: string;
    email: string;
    devices: Array<FriendDeviceDto>;
};